Improved render maintainability :godmode:
The render workflow was quite messy, manually appending children
elements and using somewhat confusing iterations. I wanted to improve
this (specially for the cases where I was manually adding an HTML
element via a string on the `innerHTML` attribute, ups). This is the
first iteration of the improvements and more will be pushed soon.